refactor(domains): extract cart/ as a dedicated subdomain
The buyer's "live cart" was scattered across orders/cart-*.ts even
though it is a distinct aggregate from a persisted Order. The cart is
mutable buyer state; an Order is the immutable artifact created in
createCheckoutOrder. Mixing them in one subdomain made it harder to
reason about boundaries.

Move five files unchanged from orders/ to a new cart/ subdomain:
  - cart-actions.ts        ('use server')
  - cart-persistence.ts    ('use server')
  - cart-dedupe.ts         (pure)
  - cart-broadcast.ts      (browser-only, BroadcastChannel + store)
  - cart-store.ts          ('use client', Zustand)

Barrel src/domains/cart/index.ts exports only server-safe modules
(actions / persistence / dedupe). cart-store.ts and cart-broadcast.ts
are client-only and must be deep-imported from client components — the
audit script enforces stores stay out of the server graph.

Drop cart-persistence from src/domains/orders/index.ts; nothing
imported the cart names from the orders barrel anyway, so this is a
zero-impact cleanup.

15 call sites migrated from @/domains/orders/cart-* to
@/domains/cart/cart-*.

No behavior change.
