+++ src/app/(buyer)/cuenta/favoritos/FavoritosClient.tsx
@@ -6,7 +6,7 @@ import Image from 'next/image'
 import { HeartIcon } from '@heroicons/react/24/solid'
 import { ShoppingCartIcon } from '@heroicons/react/24/outline'
 import { useT } from '@/i18n'
-import { useCartStore } from '@/domains/orders/cart-store'
+import { useCartStore } from '@/domains/cart/cart-store'
 import { useFavoritesStore } from '@/domains/catalog/favorites-store'
 import type { FavoriteProductItem } from '@/lib/favorites-serialization'
 import { formatPrice } from '@/lib/utils'
