+++ src/domains/orders/index.ts
@@ -1,5 +1,4 @@
 export * from './actions'
-export * from './cart-persistence'
 export * from './errors'
 export * from './checkout'
 export * from './order-event-payload'
