+++ test/features/cart-store.test.ts
@@ -1,6 +1,6 @@
 import test from 'node:test'
 import assert from 'node:assert/strict'
-import { useCartStore } from '@/domains/orders/cart-store'
+import { useCartStore } from '@/domains/cart/cart-store'
 
 function resetCart() {
   useCartStore.setState({ items: [], hasHydrated: false })
